Fix travis build and emit canonical name for capitalized proto fields in messages.

+++ test/ts_test/generate.sh
@@ -1,7 +1,10 @@
-#!/bin/bash
+#!/bin/bash -x
 
-rm -rf generated/
-mkdir -p generated/
+if [ -d generated ]
+then
+    rm -rf generated/
+fi
+mkdir generated
 
 protoc \
   --plugin=protoc-gen-ts=../../bin/protoc-gen-ts \
